moved OPTRR out of the RR.java file
